+++ kibot/registrable.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-# Copyright (c) 2020-2022 Salvador E. Tropea
-# Copyright (c) 2020-2022 Instituto Nacional de Tecnología Industrial
+# Copyright (c) 2020-2023 Salvador E. Tropea
+# Copyright (c) 2020-2023 Instituto Nacional de Tecnología Industrial
 # License: GPL-3.0
 # Project: KiBot (formerly KiPlot)
 from collections import OrderedDict
@@ -11,6 +11,12 @@ from .error import KiPlotConfigurationError
 from . import log
 
 logger = log.get_logger()
+
+
+def fname(file):
+    if file:
+        return ", while importing from `{}`".format(file)
+    return ""
 
 
 class Registrable(object):
@@ -43,10 +49,12 @@ class RegOutput(Optionable, Registrable):
         Used by BaseOutput.
         Here because it doesn't need macros. """
     _registered = {}
-    # List of defined filters
+    # Defined filters
     _def_filters = {}
-    # List of defined variants
+    # Defined variants
     _def_variants = {}
+    # Defined groups
+    _def_groups = {}
     # List of defined outputs
     _def_outputs = OrderedDict()
 
@@ -55,10 +63,12 @@ class RegOutput(Optionable, Registrable):
 
     @staticmethod
     def reset():
-        # List of defined filters
+        # Defined filters
         RegOutput._def_filters = {}
-        # List of defined variants
+        # Defined variants
         RegOutput._def_variants = {}
+        # Defined groups
+        RegOutput._def_groups = {}
         # List of defined outputs
         RegOutput._def_outputs = OrderedDict()
 
@@ -107,16 +117,29 @@ class RegOutput(Optionable, Registrable):
     @staticmethod
     def add_output(obj, file=None):
         if obj.name in RegOutput._def_outputs:
-            msg = "Output name `{}` already defined".format(obj.name)
-            if file:
-                msg += ", while importing from `{}`".format(file)
-            raise KiPlotConfigurationError(msg)
+            raise KiPlotConfigurationError("Output name `{}` already defined".format(obj.name)+fname(file))
+        if obj.name in RegOutput._def_groups:
+            raise KiPlotConfigurationError("Output name `{}` already defined as group".format(obj.name)+fname(file))
         RegOutput._def_outputs[obj.name] = obj
 
     @staticmethod
     def add_outputs(objs, file=None):
         for o in objs:
             RegOutput.add_output(o, file)
+
+    @staticmethod
+    def add_group(name, lst, file=None):
+        if name in RegOutput._def_groups:
+            raise KiPlotConfigurationError("Group name `{}` already defined".format(name)+fname(file))
+        if name in RegOutput._def_outputs:
+            raise KiPlotConfigurationError("Group name `{}` already defined as output".format(name)+fname(file))
+        RegOutput._def_groups[name] = lst
+
+    @staticmethod
+    def add_groups(objs, file=None):
+        logger.debug('Adding groups: '+str(objs))
+        for n, lst in objs.items():
+            RegOutput.add_group(n, lst, file)
 
     @staticmethod
     def get_outputs():
@@ -127,6 +150,10 @@ class RegOutput(Optionable, Registrable):
         return RegOutput._def_outputs.get(name, None)
 
     @staticmethod
+    def is_output_or_group(name):
+        return name in RegOutput._def_outputs or name in RegOutput._def_groups
+
+    @staticmethod
     def check_variant(variant):
         if variant:
             if isinstance(variant, RegVariant):
@@ -135,6 +162,23 @@ class RegOutput(Optionable, Registrable):
                 raise KiPlotConfigurationError("Unknown variant name `{}`".format(variant))
             return RegOutput.get_variant(variant)
         return None
+
+    @staticmethod
+    def solve_groups(targets, level=0):
+        """ Replaces any group by its members.
+            Returns a new list.
+            Assumes the outputs and groups are valid. """
+        new_targets = []
+        level += 1
+        if level > 20:
+            raise KiPlotConfigurationError("More than 20 levels of nested groups, possible loop")
+        for t in targets:
+            if t in RegOutput._def_outputs:
+                new_targets.append(t)
+            else:
+                # Recursive expand
+                new_targets.extend(RegOutput.solve_groups(RegOutput._def_groups[t], level))
+        return new_targets
 
 
 class RegVariant(Optionable, Registrable):